changed IDs from integer to long

+++ src/main/java/com/dgrz/cinema/entities/Category.java
@@ -11,13 +11,13 @@ public class Category {
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
 
     @Column(nullable = false)
     private String LabelCat;
 
 
-    public int getCodeCat() {
+    public long getCodeCat() {
         return id;
     }
